requirements.txt
streamlit
youtube-transcript-api
transformers
torch
openai-whisper
yt-dlp
ffmpeg-python
